Converts container packets
